requirements.txt
streamlit>=1.32.0
scikit-learn>=1.2.0
spacy>=3.7.0
pandas>=2.0.0
numpy>=1.22.0
pdfplumber>=0.11.0
python-docx>=1.1.0
plotly>=5.20.0
matplotlib>=3.8.0
sentence-transformers>=2.5.0
fastapi>=0.110.0
uvicorn>=0.29.0
mlflow>=2.11.0
motor>=3.4.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
pytest>=8.1.1
python-multipart>=0.0.9
